requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiogram-fsm-sqlitestorage==1.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiomysql==0.3.2
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.2
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
cryptography==46.0.4
easyocr==1.7.2
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
magic-filter==1.0.12
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.0
propcache==0.4.1
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
scikit-image==0.26.0
scipy==1.17.0
setuptools==82.0.0
shapely==2.1.2
six==1.17.0
SQLAlchemy==2.0.46
sqlalchemy-manager==0.0.2
sqlalchemy_orm_manager==0.0.3
sympy==1.14.0
tifffile==2026.1.28
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
yarl==1.22.0
